backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
sqlalchemy
psycopg2
supabase
google-genai
PyMuPDF
langchain-text-splitters
pydantic-settings